requirements.txt
opencv-python-headless
deta
flask
wheel
pillow
qrcode
numpy